fasthttp: switch from html/template to github.com/valyala/quicktemplate
