+++ frameworks/Go/fasthttp-mysql/setup.bat
@@ -1,2 +1,3 @@
 set GOPATH=C:\FrameworkBenchmarks\Go\fasthttp-mysql
-go run src\hello\hello.go
+go build -o server
+.\server
