+++ frameworks/Go/fasthttp-mysql/source_code
@@ -1,5 +1,4 @@
-./fasthttp-mysql/src/
-./fasthttp-mysql/src/hello
-./fasthttp-mysql/src/hello/hello.go
+./fasthttp-mysql/server.go
 ./fasthttp-mysql/templates/
-./fasthttp-mysql/templates/fortune.html
+./fasthttp-mysql/templates/fortune.qtpl
+./fasthttp-mysql/templates/fortune.qtpl.go
