+++ frameworks/Go/fasthttp-postgresql/setup.bat
@@ -1,2 +1,3 @@
 set GOPATH=C:\FrameworkBenchmarks\Go\fasthttp-postgresql
-go run src\hello\hello.go
+go build -o server
+.\server
