+++ frameworks/Go/fasthttp-postgresql/setup.sh
@@ -1,12 +1,13 @@
 #!/bin/bash
 
-sed -i 's|localhost|'"${DBHOST}"'|g' src/hello/hello.go
+sed -i 's|localhost|'"${DBHOST}"'|g' server.go
 
 fw_depends go
 
 go get -u github.com/jackc/pgx
 go get -u github.com/valyala/fasthttp
+go get -u github.com/valyala/quicktemplate/...
 
-rm -f ./hello
-go build src/hello/hello.go
-./hello &
+rm -f ./server
+go build -o server
+./server &
