--- frameworks/Go/fasthttp-postgresql/templates/fortune.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<title>Fortunes</title>
-</head>
-<body>
-<table>
-<tr><th>id</th><th>message</th></tr>
-{{range .}}
-<tr><td>{{.Id}}</td><td>{{.Message}}</td></tr>
-{{end}}
-</table>
-</body>
-</html>
